вынос типов в отдельную директорию, исправление форматирования
